@@ -1,5 +1,5 @@
 // ASN.1 JavaScript decoder
-// Copyright (c) 2008-2023 Lapo Luchini <lapo@lapo.it>
+// Copyright (c) 2008-2024 Lapo Luchini <lapo@lapo.it>
 
 // Permission to use, copy, modify, and/or distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -13,15 +13,11 @@
 // ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 // OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-(typeof define != 'undefined' ? define : function (factory) { 'use strict';
-    if (typeof module == 'object') module.exports = factory(function (name) { return require(name); });
-    else window.dom = factory(function (name) { return window[name.substring(2)]; });
-})(function (require) {
-'use strict';
+import { ASN1 } from './asn1.js';
+import { oids } from './oids.js';
+import { bindContextMenu } from './context.js';
 
 const
-    ASN1 = require('./asn1'),
-    oids = require('./oids'),
     lineLength = 80,
     contentLength = 8 * lineLength,
     DOM = {
@@ -55,7 +51,7 @@ const
         },
     };
 
-class ASN1DOM extends ASN1 {
+export class ASN1DOM extends ASN1 {
 
     toDOM(spaces) {
         spaces = spaces || '';
@@ -147,9 +143,7 @@ class ASN1DOM extends ASN1 {
                 sub.appendChild(this.sub[i].toDOM(spaces));
         }
         node.appendChild(sub);
-        head.onclick = function () {
-            node.className = (node.className == 'node collapsed') ? 'node' : 'node collapsed';
-        };
+        bindContextMenu(node);
         return node;
     }
     fakeHover(current) {
@@ -192,6 +186,7 @@ class ASN1DOM extends ASN1 {
                 this.className = 'hex';
             }
         };
+        bindContextMenu(node);
         if (root == node) {
             let lineStart = this.posStart() & 0xF;
             if (lineStart != 0) {
@@ -236,7 +231,3 @@ class ASN1DOM extends ASN1 {
     }
 
 }
-
-return ASN1DOM;
-
-});
